make sure the parent of the output file exists before opening the stream to write
fixes https://bugs.launchpad.net/prom/+bug/658929

+++ src/main/java/org/syphr/prom/ManagedProperties.java
@@ -269,6 +269,14 @@ import java.util.concurrent.locks.ReentrantReadWriteLock;
         gatekeeper.lock();
         try
         {
+            File parent = file.getParentFile();
+            if (!parent.exists() && !parent.mkdirs())
+            {
+                throw new IOException("Directory at \""
+                                      + parent.getAbsolutePath()
+                                      + "\" does not exist and cannot be created");
+            }
+            
             FileOutputStream outputStream = new FileOutputStream(file);
             try
             {
